pkg/cvo/egress: Pull HTTPS/Proxy egress into separate file
These are not just for available updates, they're also for downloading
signatures.  Placing them in a separate file makes it easier to focus
on the code that is specific to available updates.
